+++ routes/item.js
@@ -5,6 +5,7 @@ const route = express.Router();
 const fileUpload = require('express-fileupload');
 const Item = require('../db/models').itemModel;
 const User = require('../db/models').userModel;
+const authToken = require('../config/auth').authToken
 const { uploadToCloudinary, parseImage, removeFromCloudinary } = require('../config/cloudinary-config')
 
 //API handlers
@@ -17,7 +18,9 @@ route.get('/', (req, res, next) => {
         .then((item) => {
             res.status(200).send(item);
         })
-        .catch(next);
+        .catch(err => {
+            next();
+        });
 });
 
 //GET Items from Search Bar(name):                  
@@ -35,7 +38,9 @@ route.get('/search', (req, res, next) => {
         .then((item) => {
             res.send(item);
         })
-        .catch(next);
+        .catch(err => {
+            next();
+        });
 
 });
 
@@ -70,42 +75,46 @@ route.get('/:id', (req, res, next) => {
 //Middleware:
 route.use(fileUpload());
 //POST Handler:
-route.post('/', parseImage, async (req, res, next) => {
+route.post('/', authToken, parseImage, async (req, res, next) => {
     try {
-        const { body } = req;
-        let setCategories = [];
-        let index = 0;
-        for (let key of Object.keys(body)) {
-            var value = body[key];
-            if (value == 'on') {
-                setCategories[index] = key;
-                index++;
+        if (req.auth.email === req.body.userEmail) {
+            const { body } = req;
+            let setCategories = [];
+            let index = 0;
+            for (let key of Object.keys(body)) {
+                var value = body[key];
+                if (value == 'on') {
+                    setCategories[index] = key;
+                    index++;
+                }
             }
-        }
-        let imageLinks = [];
+            let imageLinks = [];
 
-        for (let encoded of req.files.encodedUri) {
-            let uploadedUrl = await uploadToCloudinary(encoded)
-            imageLinks.push({
-                url: uploadedUrl.url,
-                public_id: uploadedUrl.public_id
-            });
-        }
+            for (let encoded of req.files.encodedUri) {
+                let uploadedUrl = await uploadToCloudinary(encoded)
+                imageLinks.push({
+                    url: uploadedUrl.url,
+                    public_id: uploadedUrl.public_id
+                });
+            }
 
-        const itemBody = {
-            title: req.body.title,
-            description: req.body.description,
-            price: req.body.price,
-            userName: req.body.userName,
-            userEmail: req.body.userEmail,
-            categories: setCategories,
-            images: imageLinks
+            const itemBody = {
+                title: req.body.title,
+                description: req.body.description,
+                price: req.body.price,
+                userName: req.body.userName,
+                userEmail: req.body.userEmail,
+                categories: setCategories,
+                images: imageLinks
+            }
+            Item.create(itemBody)
+                .then((item) => {
+                    res.status(201).send(item);
+                })
+                .catch(next);
+        } else {
+            res.status(403).end();
         }
-        Item.create(itemBody)
-            .then((item) => {
-                res.status(201).send(item);
-            })
-            .catch(next);
     } catch (err) {
         next(err);
     }
@@ -113,40 +122,44 @@ route.post('/', parseImage, async (req, res, next) => {
 
 //--------------------------------------------------------------------------//
 //Send a Buy-Sell Notification 
-route.put('/notify/:id', (req, res, next) => {
+route.put('/notify/:id', authToken, (req, res, next) => {
     Item.findById(req.params.id)
         .select('userEmail userName')
         .then((item) => {
-            User.findOne({
-                email: item.userEmail,
-                notifications: {
-                    $elemMatch: {
-                        message: req.body.notification.message,
-                        userEmail: req.body.notification.userEmail
-                    }
-                }
-            })
-                .then(user => {
-                    if (user != null) {
-                        res.header("Access-Control-Allow-Origin", "*");
-                        res.status(200).send(`Already notified ${item.userName}`);
-                    } else {
-                        req.body.notification.read = false;
-                        req.body.notification._id = new mongoose.Types.ObjectId();
-
-                        const io = require('../config/socket').get();
-                        io.to(item.userEmail).emit('notification', req.body.notification)
-
-                        User.updateOne({ email: item.userEmail },
-                            { "$push": { "notifications": req.body.notification } },
-                        )
-                            .then(() => {
-                                res.status(200).send(`Notified ${item.userName}`);
-                            })
-
+            if (req.auth.email === item.userEmail) {
+                User.findOne({
+                    email: item.userEmail,
+                    notifications: {
+                        $elemMatch: {
+                            message: req.body.notification.message,
+                            userEmail: req.body.notification.userEmail
+                        }
                     }
                 })
-                .catch(next);
+                    .then(user => {
+                        if (user != null) {
+                            res.header("Access-Control-Allow-Origin", "*");
+                            res.status(200).send(`Already notified ${item.userName}`);
+                        } else {
+                            req.body.notification.read = false;
+                            req.body.notification._id = new mongoose.Types.ObjectId();
+
+                            const io = require('../config/socket').get();
+                            io.to(item.userEmail).emit('notification', req.body.notification)
+
+                            User.updateOne({ email: item.userEmail },
+                                { "$push": { "notifications": req.body.notification } },
+                            )
+                                .then(() => {
+                                    res.status(200).send(`Notified ${item.userName}`);
+                                })
+
+                        }
+                    })
+                    .catch(next);
+            } else {
+                res.status(403).end();
+            }
         })
         .catch(next);
 })
@@ -154,46 +167,54 @@ route.put('/notify/:id', (req, res, next) => {
 
 
 //Edit a posted item                               
-route.put('/:id', (req, res, next) => {
-    Item.updateOne({ _id: req.params.id }, req.body)
-        .then(
-            Item.findById(req.params.id)
-                .then((item) => {
-                    res.status(200).send(`Your Ad, ${item.name} has been updated`);
-                })
-        )
-        .catch(next);
+route.put('/:id', authToken, (req, res, next) => {
+    Item.findById(req.params.id)
+        .then((item) => {
+            if (req.auth.email === item.userEmail) {
+                Item.updateOne({ _id: req.params.id }, req.body)
+                    .then(() => {
+                        res.status(200).send(`Your Ad, ${item.name} has been updated`);
+                    })
+                    .catch(next);
+            } else {
+                res.status(403).end();
+            }
+        })
 });
 
 //-------------------------------------------------------------------------//
 
 
 //Delete a posted item                              
-route.delete('/:id', (req, res, next) => {
+route.delete('/:id', authToken, (req, res, next) => {
     Item.findById(req.params.id)
         .select('images userEmail')
         .then(item => {
-            for (let image of item.images) {
-                removeFromCloudinary(image.public_id)
-                    .catch(next);
-            };
-
-            User.updateOne({ email: item.userEmail },
-                { "$pull": { notifications: { itemId: req.params.id } } }
-            )
-                .then(() => {
-                    User.findOne({ email: item.userEmail })
-                        .select('notifications')
-                        .then(user => {
-                            Item.deleteOne({ _id: req.params.id })
-                                .then(() => {
-                                    res.status(200).send(user.notifications);
-                                })
-                                .catch(next);
-                        })
+            if (req.auth.email === item.userEmail) {
+                for (let image of item.images) {
+                    removeFromCloudinary(image.public_id)
                         .catch(next);
-                })
-                .catch(next);
+                };
+
+                User.updateOne({ email: item.userEmail },
+                    { "$pull": { notifications: { itemId: req.params.id } } }
+                )
+                    .then(() => {
+                        User.findOne({ email: item.userEmail })
+                            .select('notifications')
+                            .then(user => {
+                                Item.deleteOne({ _id: req.params.id })
+                                    .then(() => {
+                                        res.status(200).send(user.notifications);
+                                    })
+                                    .catch(next);
+                            })
+                            .catch(next);
+                    })
+                    .catch(next);
+            } else {
+                res.status(403).end();
+            }
         })
         .catch(next);
